mv separate_env_variables() in declare_arg.c and rename var_separate_key_value_op()
